feat: add shared-types package with UserJWTPayload type and update dependencies in api-gateway and auth-service
