+++ apps/api-gateway/package.json
@@ -29,6 +29,7 @@
     "@nestjs/swagger": "^11.2.0",
     "@task-mgmt/database": "workspace:*",
     "@task-mgmt/shared-config": "workspace:*",
+    "@task-mgmt/shared-types": "workspace:*",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.2",
     "drizzle-orm": "^0.44.4",
