+++ packages/shared-types/README.md
@@ -1,0 +1,13 @@
+# @task-mgmt/shared-types
+
+Shared TypeScript type definitions used across all microservices in the task management monorepo.
+
+## Usage
+
+```typescript
+import { UserJWTPayload } from '@task-mgmt/shared-types';
+```
+
+## Available Types
+
+- `UserJWTPayload` - JWT token payload structure for user authentication
